requirements.txt
backports-abc==0.5
boto==2.48.0
bz2file==0.98
certifi==2017.7.27.1
chardet==3.0.4
cycler==0.10.0
Cython==0.27.3
gensim==2.3.0
idna==2.6
matplotlib==2.0.2
nltk==3.2.5
numpy==1.13.1
pyparsing==2.2.0
python-dateutil==2.6.1
python-Levenshtein==0.12.0
pytz==2017.2
requests==2.18.4
scikit-learn==0.19.0
scipy==0.19.1
six==1.10.0
sklearn==0.0
smart-open==1.5.3
sparsesvd==0.2.2
typing==3.6.2
urllib3==1.22
